Auth__user()

+++ app/Http/Controllers/TaskController.php
@@ -4,43 +4,70 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
-use App\Models\Category;
 use App\Models\Task;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class TaskController extends Controller
 {
     public function store(StoreTaskRequest $request)
     {
-        $new_task = Task::create($request->validated());
+        $user_id = Auth::user()->id;
+
+        if (!$user_id)
+            return response()->json(["message" => "i cant access to the id"], 403);
+
+        $validated = $request->validated();
+        $validated['user_id'] = $user_id;
+
+        $new_task = Task::create($validated);
+
         return response()->json($new_task, 200);
     }
 
     public function index()
     {
-        $tasks = Task::all();
-        return  response()->json(["message" => "All Tasks", "data" => $tasks], 200);
+        $user = Auth::user();
+
+        if (!$user)
+            return response()->json(["message" => "i cant access to the user"], 403);
+
+        $tasks = $user->tasks()->get();
+        return response()->json(["message" => "All Tasks for user id : $user->id", "data" => $tasks], 200);
     }
 
     public function update(UpdateTaskRequest $request, $id)
     {
-        $newTask = Task::find($id);
+        $user_id = Auth::user()->id;
 
-        if (!$newTask)
+        if (!$user_id)
+            return response()->json(["message" => "i cant access to the id"], 403);
+
+        $updateTask = Task::find($id);
+
+        if (!$updateTask)
             return response()->json(["message" => "didnt find id " . $id], 404);
 
-        $newTask->update($request->validated());
+        if ($user_id != $updateTask->user_id)
+            return response()->json(["message" => "you are not authorized to update this task"], 403);
 
-        return response()->json([$newTask], 200);
+        $updateTask->update($request->validated());
+
+        return response()->json($updateTask, 200);
     }
 
     public function show($id)
     {
         $task = Task::find($id);
 
+        $user = Auth::user();
+
         if (!$task)
             return response()->json(["message" => "didnt find id $id"], 404);
+
+        if ($task->user()->id != $user->id)
+            return response()->json(["message" => "you are not authorized to see this task"], 403);
 
         return response()->json(["Task" => $task], 200);
     }
@@ -49,8 +76,13 @@ class TaskController extends Controller
     {
         $task = Task::find($id);
 
+        $user = Auth::user();
+
         if (!$task)
             return response()->json(["message" => "didnt find id $id"], 404);
+
+        if ($task->user()->id != $user->id)
+            return response()->json(["message" => "you are not authorized to delete this task"], 403);
 
         $task->delete();
         return response()->json(["message" => "the id $id is deleted"], 200);
@@ -58,37 +90,30 @@ class TaskController extends Controller
 
     public function deleteLow()
     {
-        $task = Task::where("priority","<=","3")->delete();
-        return response()->json(["message" => "Low priority tasks deleted",$task], 200);
+        $task = Task::where("priority", "<=", "3")->delete();
+        return response()->json(["message" => "Low priority tasks deleted", $task], 200);
     }
 
-    public function show_task_by_userId($id)
+    public function addCatToTask($taskId, Request $re)
     {
-        $user=User::find($id);
+        $task = Task::find($taskId);
 
-        if (!$user)
-            return response()->json(["message"=>"there is no user with this id "], 404);
-
-        $tasks=$user->tasks()->get();
-
-        return response()->json([$user,$tasks], 200);
-    }
-
-    public function addCatToTask($taskId,Request $re)
-    {
-        $task=Task::find($taskId);
+        $user = Auth::user();
 
         if (!$task)
-            return response()->json(["message"=>"there is no task with this id "], 404);
+            return response()->json(["message" => "there is no task with this id "], 404);
 
-        return response()->json(["message"=>"Category added to task"], 200);
+        if ($task->user()->id != $user->id)
+            return response()->json(["message" => "you are not authorized to this task"], 403);
+
+        return response()->json(["message" => "Category added to task"], 200);
     }
 
     public function show_cat_of_task($taskId)
     {
-        $task=Task::find($taskId);
+        $task = Task::find($taskId);
         if (!$task)
-            return response()->json(["message"=>"there is no task"], 200);
+            return response()->json(["message" => "there is no task"], 200);
 
         return response()->json(["Task" => $task, "categories" => $task->categories()->get()], 200);
     }
